fix bug path

+++ src/main/util/fs.js
@@ -1,26 +1,27 @@
 import fs from 'fs'
 import path from 'path'
 import CustomError from './CustomError'
+import { DEST_IMG } from './path'
 
 export const isExistsFile = (path) => {
     return fs.existsSync(path)
 }
 
-export const copyTo = (srcPath, destDir) => {
+export const copyToDestImg = (srcPath) => {
     try {
         if (!isExistsFile(srcPath)) {
             throw new CustomError('Arquivo de origem não existe')
         }
 
-        if (!isExistsFile(destDir)) {
-            fs.mkdirSync(destDir, { recursive: true })
+        if (!isExistsFile(DEST_IMG)) {
+            fs.mkdirSync(DEST_IMG, { recursive: true })
         }
 
         const fileName = path.basename(srcPath)
         const ext = path.extname(fileName)
         const baseName = path.basename(fileName, ext)
         const uniqueName = `${baseName}-${Date.now()}${ext}`
-        const destPath = path.join(destDir, uniqueName)
+        const destPath = path.join(DEST_IMG, uniqueName)
 
         fs.copyFileSync(srcPath, destPath)
 
@@ -29,9 +30,7 @@ export const copyTo = (srcPath, destDir) => {
         if (error instanceof CustomError) {
             throw error
         }
-
-        console.log(error)
-        throw new CustomError('Erro ao salvar imagem!')
+        throw new CustomError('Erro ao salvar imagem!',error)
     }
 
 }
@@ -41,6 +40,10 @@ export const updateImg = (currentImg, nextImg) => {
 
         if (!isExistsFile(nextImg)) {
             throw new CustomError('Arquivo de origem não existe')
+        }
+
+        if (!isExistsFile(currentImg)) {
+           return copyTo(nextImg)
         }
         
         fs.unlinkSync(currentImg)
@@ -61,9 +64,9 @@ export const updateImg = (currentImg, nextImg) => {
     }
 }
 
-export const deleteImg = (path) => {
+export const deleteImg = (fileName) => {
     try {
-        fs.unlinkSync(path)
+        fs.unlinkSync(path.join(DEST_IMG,fileName))
     } catch (err) {
         throw new CustomError("Erro ao deletar Imagem do produto")
     }
